// src/interfaces/bo.ts
export interface IDadosProfissionais {
  ID_DADOS_PROF: string
  ID_PROFISSAO: string
  NM_PROFISSAO: string
  NOME_EMPRESA: string
  FONE_COMERCIAL: string
  ID_RAMO_ATIVIDADE: string
  NM_RAMO_ATIVIDADE: string
  ENDERECO_COMERCIAL: IEndereco | null
}
// ...
export interface IEndereco {
  ID_ENDERECO: string
  LOGRADOURO: string
  CEP: string
  ID_MUNICIPIO: string
  MUNICIPIO: string
  ID_UF: string
  NM_UF?: string
  ID_TP_PT_REF: string
  NM_TP_PT_REF: string
  PONTO_REFERENCIA: string
  ID_TRECHO?: string
  COMPLEMENTO: string
  NUMERO: string
  BAIRRO: string
  LAT?: string | null
  LON?: string | null
}